Fix stack smashing
The function prvReportStackUsage passed a local stack variable of type
uint32_t as output parameter to xTraceKernelPortGetUnusedStack which
expected uint64_t on a 64-bit platform. This resulted in 4 byte memory
corruption.

The problem was that xTraceKernelPortGetUnusedStack uses
TraceUnsignedBaseType_t type for the output parameter which gets defined
to uint32_t on a 32-bit platform and to uint64_t on a 64-bit platform,
while the local stack variable in prvReportStackUsage was declared as
uint32_t. This commit changes the type of local stack variable in
prvReportStackUsage to TraceUnsignedBaseType_t to match the function
parameter type.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>